Fix some cancan ability and before_action permission bugs

+++ app/controllers/conference_users_controller.rb
@@ -1,7 +1,6 @@
 class ConferenceUsersController < ApplicationController
 
   before_action :require_user
-  before_action :require_admin, except: [:index]
 
   def index
     if params[:user_id]
@@ -15,6 +14,7 @@ class ConferenceUsersController < ApplicationController
 
   def create
     conference_user = ConferenceUser.new conference_user_params
+    conference_user.user_id = current_user.id unless current_user.admin? # it's all about you unless you're an admin
     conference_user.creator_id = current_user.id
     if conference_user.save
       redirect_to conference_path(conference_user.conference_id)
@@ -29,7 +29,7 @@ class ConferenceUsersController < ApplicationController
     conference_user = ConferenceUser.find(params[:id])
     if conference_user
       conference_id = conference_user.conference_id
-      conference_user.destroy
+      conference_user.destroy if conference_user.user_id == current_user.id || current_user.admin?
       redirect_to conference_path(conference_id)
     else
       render body: nil
